Added shared user list to project info page
