+++ static/components/User/UserCard/UserCard.module.css
@@ -4,6 +4,7 @@
     border: 1px grey solid;
     align-items: center;
     padding: 4px;
+    margin-bottom: 5px;
 }
 
 .avatar {
