+++ static/components/User/UsersList/UsersList.tsx
@@ -1,17 +1,16 @@
 import React from 'react';
-import { ListGroup, Image, Button } from 'react-bootstrap';
+import { ListGroup } from 'react-bootstrap';
 import { MinUserInfoT } from 'types/userTypes';
-import DefaultImage from '@assets/user.webp';
 
 import css from './UsersList.module.css';
+import { UserCard } from 'components/User/UserCard/UserCard';
 
 interface UsersListProps {
     users?: MinUserInfoT[];
-    onDelete: (username: string) => void;
 }
 
 export function UsersList(props: UsersListProps) {
-    const { users, onDelete } = props;
+    const { users } = props;
     if (!users) {
         return null;
     }
@@ -19,40 +18,12 @@ export function UsersList(props: UsersListProps) {
         <div className={css.root}>
             <ListGroup>
                 {users.map(user =>
-                    <UserRow
+                    <UserCard
                         key={user.username}
-                        user={user}
-                        onDelete={onDelete}
+                        userInfo={user}
                     />,
                 )}
             </ListGroup>
         </div>
-    );
-}
-
-interface UserRowProps {
-    user: MinUserInfoT;
-    onDelete: UsersListProps['onDelete'];
-}
-
-function UserRow(props: UserRowProps) {
-    const { user: { username, avatar }, onDelete } = props;
-    const deleteUser = React.useCallback(() => {
-        onDelete(username);
-    }, [onDelete, username]);
-    const imageSrc = avatar || DefaultImage;
-    return (
-        <ListGroup.Item key={username} className={css.userItem}>
-            <Image
-                src={imageSrc}
-                width={50}
-                height={50}
-            />
-            <div className={css.username}>{username}</div>
-            <Button
-                variant='danger'
-                className={css.removeButton}
-                onClick={deleteUser}>{'Remove'}</Button>
-        </ListGroup.Item>
     );
 }
